+++ tests/berlin/eip2929_gas_cost_increases/__init__.py
@@ -1,0 +1,4 @@
+"""
+abstract: Tests [EIP-2929: Gas cost increases for state access opcodes](https://eips.ethereum.org/EIPS/eip-2929)
+    Test cases for [EIP-2929: Gas cost increases for state access opcodes](https://eips.ethereum.org/EIPS/eip-2929).
+"""  # noqa: E501
